refactor auth packaging
