+++ packages/core/src/server/auth/auth-utils.ts
@@ -1,7 +1,7 @@
 import * as crypto from "crypto"
 import {nanoid} from "nanoid"
 import SecurePasswordLib from "secure-password"
-import {AuthenticationError} from "./errors"
+import {AuthenticationError} from "../../errors"
 
 export const hash256 = (input: string = "") =>
   crypto.createHash("sha256").update(input).digest("hex")
