+++ packages/core/src/server/auth/secure-proxy-middleware.test.ts
