+++ packages/core/src/server/auth/secure-proxy-middleware.ts
@@ -1,4 +1,4 @@
-import {Middleware, MiddlewareRequest, MiddlewareResponse} from "../types"
+import {Middleware, MiddlewareRequest, MiddlewareResponse} from "../../types"
 
 export const secureProxyMiddleware: Middleware = function (
   req: MiddlewareRequest,
