+++ packages/core/src/server/index.ts
@@ -11,10 +11,13 @@ export {
 
 export {invokeWithMiddleware} from "./invoke-with-middleware"
 
-export {passportAuth} from "./passport-adapter"
-
-export {paginate} from "./server-utils"
+export {paginate, isLocalhost} from "./server-utils"
 export type {PaginateArgs} from "./server-utils"
 
 export {resolver} from "./resolver"
 export type {AuthenticatedMiddlewareCtx} from "./resolver"
+
+export {sessionMiddleware, simpleRolesIsAuthorized, getSessionContext} from "./auth/sessions"
+export type {SimpleRolesIsAuthorized} from "./auth/sessions"
+export {passportAuth} from "./auth/passport-adapter"
+export {SecurePassword, hash256, generateToken} from "./auth/auth-utils"
