+++ packages/core/src/server/resolver.ts
@@ -1,5 +1,5 @@
 import {infer as zInfer, ZodSchema} from "zod"
-import {AuthenticatedSessionContext, SessionContext, SessionContextBase} from "../supertokens"
+import {AuthenticatedSessionContext, SessionContext, SessionContextBase} from "../auth/auth-types"
 import {Await, Ctx, EnsurePromise} from "../types"
 
 interface ResultWithContext<Result = unknown, Context = unknown> {
