+++ packages/server/src/index.ts
@@ -8,7 +8,6 @@ export type {ServerConfig} from "./config"
 export {resolveBinAsync} from "./resolve-bin-async"
 export {ManifestLoader} from "./stages/manifest"
 export * from "./rpc"
-export * from "./supertokens"
 
 // -----------------
 // For custom server
